fail cleanly if zero fast registers + no slow registers prevent specialisation
* jitter/jitter-routine.c (jitter_routine_print): Revert previous change:
make an executable routine if labels need resolving.

* jitterc/jitterc-generate.c (jitterc_emit_specializer_recognizers): If
specialisation is impossible fail cleanly at run time when the instruction
is used, instead of crashing at generation time.
(jitterc_emit_specializer): Whitespace.

* jitterc/jitterc-vm.h (jitterc_vm_register_class_set_fast_register_no): Change
parameter type, accepting a signed value...
* jitterc/jitterc-vm.c (jitterc_vm_register_class_set_fast_register_no): But
check the sign here.  This catches invalid values in the Jitter specification.

+++ jitterc/jitterc-vm.h
@@ -129,7 +129,7 @@ jitterc_vm_register_class_set_c_initial_value (struct jitterc_register_class *rc
                                                const char *c_initial_value);
 void
 jitterc_vm_register_class_set_fast_register_no (struct jitterc_register_class
-                                                *rc, size_t fast_register_no);
+                                                *rc, int fast_register_no);
 void
 jitterc_vm_register_class_set_use_slow_registers (struct jitterc_register_class
                                                   *rc, int use_slow_registers);
